Add files via upload
Given an integer product, find the smallest positive (i.e. greater than 0) integer the product of whose digits is equal to product. If there is no such integer, return -1 instead.

Example

For product = 12, the output should be solution(product) = 26;
For product = 19, the output should be solution(product) = -1.
For product = 33, the output should be solution(product) = -1.